# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Students of Islamic Law Forum API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List members (members only)
      parameters:
        - in: query
          name: gender
          schema:
            type: string
            enum: [male, female]
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401":
          description: Authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile (members only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /chat/groups/{id}:
    get:
      operationId: getChatGroup
      tags: [chat]
      summary: Get chat group details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatGroup" }
        "403":
          description: Forbidden (gender restricted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /chat/groups/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Messages in a chat group
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: postChatMessage
      tags: [chat]
      summary: Post a chat message
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChatMessageBody" }
      responses:
        "201":
          description: Posted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /storage/uploads:
    post:
      operationId: uploadFile
      tags: [storage]
      summary: Upload an image or video file (multipart/form-data)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResponse" }
        "400":
          description: Invalid file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Gender:
      type: string
      enum: [male, female]
    User:
      type: object
      required: [id, username, displayName, gender, isAdmin, isMainAdmin, isActive, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email:
          type: ["string", "null"]
          format: email
        gender: { $ref: "#/components/schemas/Gender" }
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin: { type: boolean }
        isMainAdmin: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 60
        email:
          type: string
          format: email
          maxLength: 254
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...
    ChatGroup:
      type: object
      required: [id, name, gender, memberCount, lastMessageAt, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description:
          type: ["string", "null"]
        gender: { $ref: "#/components/schemas/Gender" }
        coverImageUrl:
          type: ["string", "null"]
        memberCount: { type: integer }
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
    ChatMessage:
      type: object
      required: [id, groupId, userId, author, content, createdAt]
      properties:
        id: { type: integer }
        groupId: { type: integer }
        userId: { type: integer }
        author: { $ref: "#/components/schemas/User" }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    CreateChatMessageBody:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
...
    UploadResponse:
      type: object
      required: [objectPath, url, kind, contentType]
      properties:
        objectPath: { type: string }
        url: { type: string }
        kind:
          type: string
          enum: [image, video]
        contentType: { type: string }
...