# .migration-backup/lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Students of Islamic Law Forum API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileBody {
  /**
   * @minLength 2
   * @maxLength 60
   */
  displayName?: string;
  /** @maxLength 254 */
  email?: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}